Fix bugs

+++ backend/channel/BucketsChannel.js
@@ -6,28 +6,23 @@ const {spawn} = require('child_process');
 
 class BucketsChannel extends Channel {
 
-    async getBuckets(profile) {
-
-        const s3 = S3Helper.getS3(profile);
-
-        const response = await s3.send(new ListBucketsCommand({}));
-
-        return response.Buckets;
-    }
-
     async getBucketsWithAcl(profile) {
 
         const s3 = S3Helper.getS3(profile);
 
         const response = await s3.send(new ListBucketsCommand({}));
 
-        return await Promise.all(response.Buckets.map(async (bucket) => {
+        let buckets = await Promise.all(response.Buckets.map(async (bucket) => {
 
             bucket.IsPublic = await S3Helper.isPublicBucket(s3, bucket.Name);
 
             return bucket;
 
         }));
+
+        return buckets.filter(function (bucket) {
+            return bucket.IsPublic !== null
+        });
 
     }
 
@@ -41,12 +36,77 @@ class BucketsChannel extends Channel {
             throw {Code: "ExistBucket"}
         }
 
+        /* use rclone instead
+
         const data = await s3.send(
             new CreateBucketCommand({
                 Bucket: bucketName,
+                // CreateBucketConfiguration: {
+                //     LocationConstraint: ''
+                // },
                 ACL: acl // 'private' | 'public-read'
             })
         );
+
+         */
+
+        let temp_url = profile
+            .endpoint_url
+            .trim()
+            .replace("https://s3.", "")
+            .replace("http://s3.", "");
+
+        let region = temp_url.substring(0, temp_url.indexOf("."));
+
+        let config = `
+[s3]
+type = s3
+provider = Other
+access_key_id = ${profile.access_key}
+secret_access_key = ${profile.secret_key}
+${profile.provider != "arvan" ? `region = ${region}` : ``}
+endpoint = ${profile.endpoint_url}`;
+
+        await fs.writeFile("rclone/rclone-config.conf", config);
+
+        if(GlobalData.RClone){
+            GlobalData.RClone.kill();
+        }
+
+        return await new Promise((resolve, reject) => {
+
+            GlobalData.RClone = spawn('rclone\\rclone', ['mkdir', 's3:' + bucketName, '--s3-acl', acl], {
+                env: {
+                    ...process.env,
+                    RCLONE_CONFIG: "rclone\\rclone-config.conf"
+                }
+            });
+
+            GlobalData.RClone.stdout.on('data', (data) => {
+                console.log(`stdout: ${data}`);
+                reject(`${data}`);
+            });
+
+            GlobalData.RClone.stderr.on('data', (data) => {
+                console.log(`stderr: ${data}`);
+                reject(`${data}`);
+
+            });
+
+            GlobalData.RClone.on('close', async (code) => {
+
+                let exist = await S3Helper.checkExistBucket(s3, bucketName);
+
+                if (exist) {
+                    resolve();
+                }
+                else{
+                    reject("Not create!")
+                }
+
+            });
+
+        });
 
     }
 
@@ -55,6 +115,60 @@ class BucketsChannel extends Channel {
         const s3 = S3Helper.getS3(profile);
 
         await s3.send(new DeleteBucketCommand({Bucket: bucketName}));
+
+    }
+
+    async emptyBucket(profile, bucketName) {
+
+        let temp_url = profile
+            .endpoint_url
+            .trim()
+            .replace("https://s3.", "")
+            .replace("http://s3.", "");
+
+        let region = temp_url.substring(0, temp_url.indexOf("."));
+
+        let config = `
+[s3]
+type = s3
+provider = Other
+access_key_id = ${profile.access_key}
+secret_access_key = ${profile.secret_key}
+region = ${region}
+endpoint = ${profile.endpoint_url}
+storage_class = STANDARD`;
+
+        await fs.writeFile("rclone/rclone-config.conf", config);
+
+        if(GlobalData.RClone){
+            GlobalData.RClone.kill();
+        }
+
+        return await new Promise((resolve, reject) => {
+
+            GlobalData.RClone = spawn('rclone\\rclone', ['delete', 's3:' + bucketName], {
+                env: {
+                    ...process.env,
+                    RCLONE_CONFIG: "rclone\\rclone-config.conf"
+                }
+            });
+
+            // GlobalData.RClone.stdout.on('data', (data) => {
+            //     console.log(`stdout: ${data}`);
+            //     resolve();
+            // });
+
+            GlobalData.RClone.stderr.on('data', (data) => {
+                console.log(`stderr: ${data}`);
+                reject(`${data}`);
+
+            });
+
+            GlobalData.RClone.on('close', (code) => {
+                resolve();
+            });
+
+        });
 
     }
 
@@ -78,6 +192,7 @@ class BucketsChannel extends Channel {
 
         const self = this;
         const maxSingleUploadSize = 5 * 1024 * 1024;
+        const sameProfile = (sourceProfile.access_key === destProfile.access_key && sourceProfile.endpoint_url === destProfile.endpoint_url);
 
         const sourceS3 = S3Helper.getS3(sourceProfile);
         const destS3 = S3Helper.getS3(destProfile);
@@ -140,68 +255,81 @@ class BucketsChannel extends Channel {
             }
 
             sourceObjects[i].IsPublic = await S3Helper.isPublicObject(sourceS3, sourceBucketName, sourceObjects[i].Key);
-            sourceObjects[i].TempPath = await S3Helper.downloadObjectInTemp(sourceS3, sourceBucketName, sourceObjects[i].Key);
 
-            objectPercent = 30;
+            if(sameProfile){
 
-            progress();
-
-            if (sourceObjects[i].Size < maxSingleUploadSize) {
-
-                await S3Helper.putObjectInBucket(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic)
-                fs.remove(sourceObjects[i].TempPath);
+                await S3Helper.copyObject(sourceS3, sourceBucketName, destBucketName, sourceObjects[i].Key, sourceObjects[i].IsPublic);
 
                 objectPercent = 100;
 
                 progress();
             }
-            else {
+            else{
 
-                await new Promise((resolve, reject) => {
+                sourceObjects[i].TempPath = await S3Helper.downloadObjectInTemp(sourceS3, sourceBucketName, sourceObjects[i].Key);
 
-                    const mPart = S3Helper.putObjectInBucketMultiPart(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic);
+                objectPercent = 30;
 
-                    mPart.on("error", async (e) => {
+                progress();
 
-                        await fs.remove(sourceObjects[i].TempPath);
+                if (sourceObjects[i].Size < maxSingleUploadSize) {
 
-                        reject(e);
+                    await S3Helper.putObjectInBucket(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic)
+                    fs.remove(sourceObjects[i].TempPath);
 
+                    objectPercent = 100;
+
+                    progress();
+                }
+                else {
+
+                    await new Promise((resolve, reject) => {
+
+                        const mPart = S3Helper.putObjectInBucketMultiPart(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic);
+
+                        mPart.on("error", async (e) => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            reject(e);
+
+                        });
+
+                        mPart.on("abort", async () => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            self.sendTrigger('copyBucket@abort');
+
+                            GlobalData.AppInProcess = false;
+
+                            reject();
+
+
+                        });
+
+                        mPart.on("end", async (response) => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            objectPercent = 100;
+
+                            progress();
+
+                            resolve(response);
+
+                        });
+
+                        mPart.on("progress", (doneCount, totalCount) => {
+
+                            objectPercent = 30 + (doneCount * 70 / totalCount);
+
+                            progress();
+
+                        });
                     });
 
-                    mPart.on("abort", async () => {
-
-                        await fs.remove(sourceObjects[i].TempPath);
-
-                        self.sendTrigger('copyBucket@abort');
-
-                        GlobalData.AppInProcess = false;
-
-                        reject();
-
-
-                    });
-
-                    mPart.on("end", async (response) => {
-
-                        await fs.remove(sourceObjects[i].TempPath);
-
-                        objectPercent = 100;
-
-                        progress();
-
-                        resolve(response);
-
-                    });
-
-                    mPart.on("progress", (doneCount, totalCount) => {
-
-                        objectPercent = 30 + (doneCount * 70 / totalCount);
-
-                        progress();
-
-                    });
-                });
+                }
 
             }
 
@@ -220,6 +348,7 @@ class BucketsChannel extends Channel {
 
         const self = this;
         const maxSingleUploadSize = 5 * 1024 * 1024;
+        const sameProfile = (sourceProfile.access_key === destProfile.access_key && sourceProfile.endpoint_url === destProfile.endpoint_url);
 
         const sourceS3 = S3Helper.getS3(sourceProfile);
         const destS3 = S3Helper.getS3(destProfile);
@@ -263,68 +392,82 @@ class BucketsChannel extends Channel {
             progress();
 
             sourceObjects[i].IsPublic = await S3Helper.isPublicObject(sourceS3, sourceBucketName, sourceObjects[i].Key);
-            sourceObjects[i].TempPath = await S3Helper.downloadObjectInTemp(sourceS3, sourceBucketName, sourceObjects[i].Key);
 
-            objectPercent = 30;
+            if(sameProfile){
 
-            progress();
-
-            if (sourceObjects[i].Size < maxSingleUploadSize) {
-
-                await S3Helper.putObjectInBucket(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic)
-                fs.remove(sourceObjects[i].TempPath);
+                await S3Helper.copyObject(sourceS3, sourceBucketName, destBucketName, sourceObjects[i].Key, sourceObjects[i].IsPublic);
 
                 objectPercent = 100;
 
                 progress();
             }
-            else {
+            else{
 
-                await new Promise((resolve, reject) => {
 
-                    const mPart = S3Helper.putObjectInBucketMultiPart(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic);
+                sourceObjects[i].TempPath = await S3Helper.downloadObjectInTemp(sourceS3, sourceBucketName, sourceObjects[i].Key);
 
-                    mPart.on("error", async (e) => {
+                objectPercent = 30;
 
-                        await fs.remove(sourceObjects[i].TempPath);
+                progress();
 
-                        reject(e);
+                if (sourceObjects[i].Size < maxSingleUploadSize) {
 
+                    await S3Helper.putObjectInBucket(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic)
+                    fs.remove(sourceObjects[i].TempPath);
+
+                    objectPercent = 100;
+
+                    progress();
+                }
+                else {
+
+                    await new Promise((resolve, reject) => {
+
+                        const mPart = S3Helper.putObjectInBucketMultiPart(destS3, destBucketName, sourceObjects[i].TempPath, sourceObjects[i].Key, sourceObjects[i].IsPublic);
+
+                        mPart.on("error", async (e) => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            reject(e);
+
+                        });
+
+                        mPart.on("abort", async () => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            self.sendTrigger('copyBucket@abort');
+
+                            GlobalData.AppInProcess = false;
+
+                            reject();
+
+
+                        });
+
+                        mPart.on("end", async (response) => {
+
+                            await fs.remove(sourceObjects[i].TempPath);
+
+                            objectPercent = 100;
+
+                            progress();
+
+                            resolve(response);
+
+                        });
+
+                        mPart.on("progress", (doneCount, totalCount) => {
+
+                            objectPercent = 30 + (doneCount * 70 / totalCount);
+
+                            progress();
+
+                        });
                     });
 
-                    mPart.on("abort", async () => {
-
-                        await fs.remove(sourceObjects[i].TempPath);
-
-                        self.sendTrigger('copyBucket@abort');
-
-                        GlobalData.AppInProcess = false;
-
-                        reject();
-
-
-                    });
-
-                    mPart.on("end", async (response) => {
-
-                        await fs.remove(sourceObjects[i].TempPath);
-
-                        objectPercent = 100;
-
-                        progress();
-
-                        resolve(response);
-
-                    });
-
-                    mPart.on("progress", (doneCount, totalCount) => {
-
-                        objectPercent = 30 + (doneCount * 70 / totalCount);
-
-                        progress();
-
-                    });
-                });
+                }
 
             }
